# backend/functions/openapi.yaml
...
openapi: 3.0.3
info:
  title: Sparkling REST API
  version: 1.0.0
  description: >-
    Sparkling platform backend (SRS v1.0). All JSON, snake_case, ISO-8601 UTC timestamps, money in integer ZAR cents. Mutating routes accept `Idempotency-Key` (or `client_op_id` in the body); replays
    return the cached first response. Every response carries `X-Correlation-Id`. Clients send `X-Client-App` and `X-Client-Version`.
  contact:
    name: Sparkling engineering
    email: dev@sparkling.co.za
  license:
    name: Proprietary
    url: https://sparkling.co.za/legal
servers:
  - url: https://europe-west1-sparkling-4e89d.cloudfunctions.net/api/v1
    description: dev (Cloud Functions 2nd gen)
  - url: http://127.0.0.1:5001/sparkling-4e89d/europe-west1/api/v1
    description: Firebase emulator
...
security:
  - bearerAuth: []
paths:
...
  /outlets:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Outlet'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Catalogue
      operationId: listOutlets
      summary: Active outlets (optionally sorted by distance)
      parameters:
        - name: lat
          in: query
          required: false
          schema:
            type: number
        - name: lng
          in: query
          required: false
          schema:
            type: number
  /outlets/{id}/services:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/OutletServiceOffer'
                        - type: object
                          properties:
                            membership:
                              type: object
                              properties:
                                membership_id:
                                  type: string
                                  format: uuid
                                plan_code:
                                  type: string
                                plan_name:
                                  type: string
                                benefit:
                                  type: string
                                  enum:
                                    - included
                                    - discount
                                  nullable: true
                                entitlement_id:
                                  type: string
                                  format: uuid
                                  nullable: true
                                entitlement_code:
                                  type: string
                                  nullable: true
                                remaining_after:
                                  type: integer
                                  nullable: true
                                period_end:
                                  type: string
                                  format: date-time
                                  nullable: true
                              nullable: true
                  groups:
                    type: array
                    items:
                      type: string
                      enum:
                        - Car Wash Options
                        - Combinations
                        - Auto Body Repair
                  tier:
                    type: string
                    enum:
                      - silver
                      - gold
                      - platinum
                      - black
                  vehicle_size:
                    type: string
                    enum:
                      - small
                      - large
                      - bike
                    nullable: true
                  membership:
                    type: object
                    properties:
                      plan_code:
                        type: string
                      plan_name:
                        type: string
                      status:
                        type: string
                        enum:
                          - pending
                          - active
                          - past_due
                          - cancelled
                          - expired
                    nullable: true
                required:
                  - data
                  - groups
                  - tier
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Catalogue
      operationId: listOutletServices
      summary: >-
        Offers at an outlet: per-size prices, composition, add-ons, grouped (with ?vehicle_size: price_cents/total resolved for that size and the caller’s membership — covered services come back at R0
        with "Included in Gold · 3 of 4 left")
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: vehicle_size
          in: query
          required: false
          schema:
            type: string
            enum:
              - small
              - large
              - bike
  /availability:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Slot'
                  outlet_id:
                    type: string
                    format: uuid
                  service_id:
                    type: string
                    format: uuid
                  date:
                    type: string
                    format: date
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Catalogue
      operationId: getAvailability
      summary: Bookable slots for a service at an outlet on a date
      parameters:
        - name: outlet_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: service_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
...
  /loyalty/account:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  account:
                    $ref: '#/components/schemas/LoyaltyAccount'
                  tier_config:
                    type: object
                    allOf:
                      - $ref: '#/components/schemas/LoyaltyTier'
                    nullable: true
                  tiers:
                    type: array
                    items:
                      $ref: '#/components/schemas/LoyaltyTier'
                  next_tier:
                    type: object
                    properties:
                      tier:
                        type: string
                        enum:
                          - silver
                          - gold
                          - platinum
                          - black
                      name:
                        type: string
                      points_needed:
                        type: integer
                    nullable: true
                  published_version:
                    type: integer
                    nullable: true
                  rules:
                    type: object
                    allOf:
                      - $ref: '#/components/schemas/LoyaltyRules'
                    nullable: true
                  membership:
                    type: object
                    allOf:
                      - $ref: '#/components/schemas/MembershipBrief'
                    nullable: true
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Loyalty
      operationId: getLoyaltyAccount
      summary: Account, tier config (tier = membership plan; next_tier is informational only) and membership summary
  /loyalty/rewards:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reward'
                  balance_points:
                    type: integer
                  tier:
                    type: string
                    enum:
                      - silver
                      - gold
                      - platinum
                      - black
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Loyalty
      operationId: listRewards
      summary: Rewards with eligibility by tier
...
  /loyalty/ledger:
    get:
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LedgerEntry'
                  next_cursor:
                    type: string
                    nullable: true
                required:
                  - data
                  - next_cursor
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Loyalty
      operationId: getLoyaltyLedger
      summary: Ledger entries
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            example: slot_start:desc
  /loyalty/rewards/{id}/redeem:
    post:
      responses:
        '201':
          description: Redeemed (or replayed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Redemption'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '410':
          $ref: '#/components/responses/Gone'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      tags:
        - Loyalty
      operationId: redeemReward
      summary: Redeem (idempotent by Idempotency-Key)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  responses:
    ValidationError:
      description: validation_error (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthenticated:
      description: unauthenticated (401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: forbidden (403)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: not_found (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: conflict / invalid_transition (409)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Gone:
      description: gone (410) — expired quote link or quotation past valid_until
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimited:
      description: rate_limited (429)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Internal:
      description: internal (500) — never includes stack traces
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - unauthenticated
                - forbidden
                - not_found
                - validation_error
                - conflict
                - invalid_transition
                - invalid_otp
                - gone
                - rate_limited
                - internal
            message:
              type: string
            details:
              type: array
              items: {}
            correlation_id:
              type: string
          required:
            - code
            - message
            - correlation_id
      required:
        - error
...
    Outlet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        address_line:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        country:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        timezone:
          type: string
        opening_hours:
          type: object
          additionalProperties: true
        slot_minutes:
          type: integer
        bay_count:
          type: integer
        rating:
          type: number
          nullable: true
        is_active:
          type: boolean
        distance_km:
          type: number
          description: Present when ?lat&lng supplied
...
    OutletServiceOffer:
      type: object
      properties:
        service_id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
          description: Outlet display_name ?? service name
        description:
          type: string
          nullable: true
        group_name:
          type: string
          enum:
            - Car Wash Options
            - Combinations
            - Auto Body Repair
        category:
          type: string
          enum:
            - car_wash
            - auto_body
        duration_minutes:
          type: integer
        icon:
          type: string
          nullable: true
        pricing_mode:
          type: string
          enum:
            - from
            - fixed
            - by_quote
        vat_mode:
          type: string
          enum:
            - incl
            - excl
        price_small_cents:
          type: integer
          nullable: true
        price_large_cents:
          type: integer
          nullable: true
        price_general_cents:
          type: integer
          nullable: true
        price_from_cents:
          type: integer
          description: Lowest non-null price; null when by-quote
          nullable: true
        price_for:
          type: object
          properties:
            small:
              type: integer
              nullable: true
            large:
              type: integer
              nullable: true
            bike:
              type: integer
              nullable: true
          required:
            - small
            - large
            - bike
        price_label:
          type: string
          example: From R 150
        is_addon:
          type: boolean
        addon_group_name:
          type: string
          enum:
            - Car Wash Options
            - Combinations
            - Auto Body Repair
          nullable: true
        includes:
          type: array
          items:
            type: object
            properties:
              service_id:
                type: string
                format: uuid
              code:
                type: string
              name:
                type: string
              quantity:
                type: integer
        included_in:
          type: array
          items:
            type: string
            format: uuid
        components_source:
          type: string
          enum:
            - outlet
            - global
            - none
        is_available:
          type: boolean
        sort_order:
          type: integer
        notes:
          type: string
          nullable: true
        points_estimate:
          type: integer
        price_cents:
          type: integer
          description: 'Only with ?vehicle_size: price resolved for that size (base, before discount)'
          nullable: true
        discount_cents:
          type: integer
        vat_cents:
          type: integer
        total_cents:
          type: integer
          nullable: true
        discount_label:
          type: string
          nullable: true
      required:
        - service_id
        - code
        - name
        - group_name
        - category
        - pricing_mode
        - vat_mode
        - price_from_cents
        - price_for
        - price_label
        - is_addon
        - includes
        - included_in
        - components_source
        - is_available
        - sort_order
        - points_estimate
...
    Slot:
      type: object
      properties:
        slot_start:
          type: string
          format: date-time
        slot_end:
          type: string
          format: date-time
        capacity:
          type: integer
        booked:
          type: integer
        available:
          type: boolean
...
    LoyaltyTier:
      type: object
      properties:
        tier:
          type: string
          enum:
            - silver
            - gold
            - platinum
            - black
        name:
          type: string
        min_points:
          type: integer
        max_points:
          type: integer
          nullable: true
        earn_multiplier:
          type: number
        discount_pct:
          type: number
      required:
        - tier
        - name
        - min_points
        - earn_multiplier
        - discount_pct
    LoyaltyRules:
      type: object
      properties:
        points_per_rand:
          type: number
        award_on:
          type: string
        idempotent_award:
          type: boolean
        expiry_months:
          type: integer
          nullable: true
        birthday_bonus:
          type: object
          additionalProperties: true
        referral_bonus:
          type: object
          additionalProperties: true
      required:
        - points_per_rand
      additionalProperties: true
...
    LoyaltyAccount:
      type: object
      properties:
        customer_id:
          type: string
        tier:
          type: string
          enum:
            - silver
            - gold
            - platinum
            - black
        balance_points:
          type: integer
          description: Computed from the ledger (authoritative)
        lifetime_points:
          type: integer
        tier_since:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
    LedgerEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
        delta:
          type: integer
        type:
          type: string
          enum:
            - earn
            - redeem
            - adjust
            - expire
            - bonus
        source_type:
          type: string
          nullable: true
        source_id:
          type: string
          format: uuid
          nullable: true
        reference:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        idempotency_key:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    Reward:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        points_cost:
          type: integer
        min_tier:
          type: string
          enum:
            - silver
            - gold
            - platinum
            - black
        is_active:
          type: boolean
        sort_order:
          type: integer
        eligible:
          type: boolean
        affordable:
          type: boolean
    Redemption:
      type: object
      properties:
        ledger_id:
          type: string
          format: uuid
        redemption:
          type: object
          properties:
            id:
              type: string
              format: uuid
            code:
              type: string
              example: RW-1182
            status:
              type: string
            reward_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
        balance_after:
          type: integer
...
    Allowance:
      type: object
      properties:
        entitlement_id:
          type: string
          format: uuid
        entitlement_code:
          type: string
          example: G1
        group_code:
          type: string
          example: washes
        label:
          type: string
        quantity:
          type: integer
        used:
          type: integer
        remaining:
          type: integer
        period:
          type: string
          enum:
            - month
            - year
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
      required:
        - entitlement_id
        - entitlement_code
        - group_code
        - label
        - quantity
        - used
        - remaining
        - period
        - period_start
        - period_end
    MembershipBrief:
      type: object
      properties:
        membership_id:
          type: string
          format: uuid
        plan_code:
          type: string
        plan_name:
          type: string
        tier:
          type: string
          enum:
            - silver
            - gold
            - platinum
            - black
        status:
          type: string
          enum:
            - pending
            - active
            - past_due
            - cancelled
            - expired
        period_end:
          type: string
          format: date-time
          nullable: true
        cancel_at_period_end:
          type: boolean
        allowances:
          type: array
          items:
            $ref: '#/components/schemas/Allowance'
      required:
        - membership_id
        - plan_code
        - plan_name
        - tier
        - status
        - allowances
...